requirements.txt
paho-mqtt==1.6.1
redis==5.0.1
pandas==2.1.4
numpy==1.26.2
fastapi==0.108.0
uvicorn[standard]==0.25.0
websockets==12.0
python-dotenv==1.0.0